requirements-dev.txt
Flask-Migrate==2.5.3
Flask-Script==2.0.6
autopep8==1.5.1
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0